+++ src/main/java/com/mobble/mobbleserver/domain/club/club/controller/ClubController.java
@@ -1,8 +1,8 @@
-package com.mobble.mobbleserver.domain.club.controller;
+package com.mobble.mobbleserver.domain.club.club.controller;
 
-import com.mobble.mobbleserver.domain.club.dto.request.ClubRequestDto;
-import com.mobble.mobbleserver.domain.club.dto.response.ClubResponseDto;
-import com.mobble.mobbleserver.domain.club.service.ClubService;
+import com.mobble.mobbleserver.domain.club.club.dto.request.ClubRequestDto;
+import com.mobble.mobbleserver.domain.club.club.dto.response.ClubResponseDto;
+import com.mobble.mobbleserver.domain.club.club.service.ClubService;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Positive;
 import lombok.RequiredArgsConstructor;
