+++ src/main/java/com/mobble/mobbleserver/domain/club/club/repository/ClubQueryRepository.java
@@ -1,6 +1,6 @@
-package com.mobble.mobbleserver.domain.club.repository;
+package com.mobble.mobbleserver.domain.club.club.repository;
 
-import com.mobble.mobbleserver.domain.club.repository.dto.ClubLikeInfoDto;
+import com.mobble.mobbleserver.domain.club.club.repository.dto.ClubLikeInfoDto;
 
 public interface ClubQueryRepository {
