+++ src/main/java/com/mobble/mobbleserver/domain/club/club/repository/ClubRepository.java
@@ -1,6 +1,6 @@
-package com.mobble.mobbleserver.domain.club.repository;
+package com.mobble.mobbleserver.domain.club.club.repository;
 
-import com.mobble.mobbleserver.domain.club.entity.Club;
+import com.mobble.mobbleserver.domain.club.club.entity.Club;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ClubRepository extends JpaRepository<Club, Long>, ClubQueryRepository {
