+++ src/main/java/com/mobble/mobbleserver/domain/club/clubAgeGroup/repository/ClubAgeGroupRepository.java
@@ -1,7 +1,7 @@
-package com.mobble.mobbleserver.domain.clubAgeGroup.repository;
+package com.mobble.mobbleserver.domain.club.clubAgeGroup.repository;
 
-import com.mobble.mobbleserver.domain.club.entity.Club;
-import com.mobble.mobbleserver.domain.clubAgeGroup.entity.ClubAgeGroup;
+import com.mobble.mobbleserver.domain.club.club.entity.Club;
+import com.mobble.mobbleserver.domain.club.clubAgeGroup.entity.ClubAgeGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
