+++ src/main/java/com/mobble/mobbleserver/domain/club/clubAgeGroup/repository/ClubAgeGroupRepositoryImpl.java
@@ -1,15 +1,14 @@
-package com.mobble.mobbleserver.domain.clubAgeGroup.repository;
+package com.mobble.mobbleserver.domain.club.clubAgeGroup.repository;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
-import static com.mobble.mobbleserver.domain.clubAgeGroup.entity.QClubAgeGroup.clubAgeGroup;
+import static com.mobble.mobbleserver.domain.club.clubAgeGroup.entity.QClubAgeGroup.clubAgeGroup;
 
 @Repository
 @RequiredArgsConstructor
 public class ClubAgeGroupRepositoryImpl implements ClubAgeGroupQueryDslRepository{
-
     private final JPAQueryFactory queryFactory;
 
     @Override
