+++ src/main/java/com/mobble/mobbleserver/domain/like/clubLike/entity/ClubLike.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.domain.like.clubLike.entity;
 
-import com.mobble.mobbleserver.domain.club.entity.Club;
+import com.mobble.mobbleserver.domain.club.club.entity.Club;
 import com.mobble.mobbleserver.domain.member.entity.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
