+++ src/test/java/com/mobble/mobbleserver/support/fixture/club/ClubTestFixture.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.support.fixture.club;
 
-import com.mobble.mobbleserver.domain.club.entity.Club;
+import com.mobble.mobbleserver.domain.club.club.entity.Club;
 import com.mobble.mobbleserver.domain.clubCategory.entity.ClubCategory;
 
 public class ClubTestFixture {
